+++ wasm-util/detectOther.ts
@@ -1,4 +1,0 @@
-
-import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot@0.2.45/wasm-util/packageInfo.ts';
-
-export default [utilInfo];